requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.5
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonrpcclient==4.0.3
MarkupSafe==2.1.3
MarkupSafe==2.1.3
protobuf==4.25.1
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
